requirements.txt
pandas
numpy
opencv-python
scikit-learn
joblib
streamlit
Pillow
